Adapted to new metadata
git-svn-id: file:///home/roberto/Git/svn@7134 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: e2a2d80284b22c347a1f43769c01944a8cccbebe

+++ trunk/xmipp/libraries/data/program.h
@@ -245,7 +245,7 @@ protected:
     virtual void readParams();
     virtual void preProcess();
     virtual void postProcess();
-    virtual void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId) = 0;
+    virtual void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId) = 0;
     virtual void show();
     /** Do some stuff before starting processing
      * in a parallel environment usually this only be executed
